+++ R/app_01_assign_server.R
@@ -33,10 +33,23 @@ ghqc_assign_server <- function(id, remote, root_dir, checklists, org, repo, memb
     rootFolder = root_dir_reactive,
     search = FALSE,
     pattern = exclude_patterns(),
-    all.files = FALSE
+    all.files = FALSE,
+    output_id = "treeNavigator"
   )
 
   moduleServer(id, function(input, output, session) {
+
+    # This section ensures that when an error occurs, the app stops
+    # When an error occurs, the session ends. The other instance of this is when
+    # the user clicks reset.
+    # The logic here prevents the app from stopping when reset is clicked
+    reset_triggered <- reactiveVal(FALSE)
+    session$onSessionEnded(function() {
+      if (!isTRUE(isolate(reset_triggered()))) {
+        stopApp()
+      }
+    })
+
     ns <- session$ns
 
     if (length(milestone_list) == 0) {
@@ -111,16 +124,6 @@ ghqc_assign_server <- function(id, remote, root_dir, checklists, org, repo, memb
             width = "100%"
           )
         ),
-        selectizeInput(
-          ns("assignees"),
-          "Select Assignee(s)",
-          choices = "No assignee",
-          multiple = TRUE,
-          width = "100%",
-          options = list(
-            closeAfterSelect = TRUE
-          )
-        ),
         div(
           style = "font-family: \"Helvetica Neue\", Helvetica, Arial, sans-serif !important; font-weight: bold;",
           "Select File(s) for QC"
@@ -138,6 +141,7 @@ ghqc_assign_server <- function(id, remote, root_dir, checklists, org, repo, memb
         iv$add_rule("milestone_existing", shinyvalidate::sv_required())
       }
     })
+
 
     observe({
       req(org, members)
@@ -165,7 +169,7 @@ return "<div><strong>" + escape(item.username) + "</div>"
 }'
           )
         )
-      )
+      ) # updateSelectizeInput
     })
 
     observe({
@@ -191,10 +195,16 @@ return "<div><strong>" + escape(item.username) + "</div>"
       req(root_dir, selected_items())
       tryCatch(
         {
-          file_data <- extract_file_data(input, selected_items())
+          relevant_files_list <- tryCatch({
+            relevant_files()
+          }, error = function(e){
+            NULL
+          })
+          file_data <- extract_file_data(input, selected_items(), relevant_files_list)
         },
         error = function(e) {
           error(.le$logger, glue::glue("There was an error extracting file data from {selected_items()}:{e$message}"))
+          stopApp()
           rlang::abort(e$message)
         }
       )
@@ -212,26 +222,74 @@ return "<div><strong>" + escape(item.username) + "</div>"
 
     })
 
-    output$main_panel_dynamic <- renderUI({
-      validate(need(length(selected_items()) > 0, "No files selected"))
-      w_load_items$show()
+    relevant_files <- reactiveVal(list())
 
-      log_string <- glue::glue_collapse(selected_items(), sep = ", ")
-      debug(.le$logger, glue::glue("Files selected for QC: {log_string}"))
-
-      list <- render_selected_list(input, ns, iv, items = selected_items(), checklist_choices = checklists)
-      isolate_rendered_list(input, session, selected_items())
-
-      session$sendCustomMessage("adjust_grid", id) # finds the width of the files and adjusts grid column spacing based on values
-      return(list)
+    output$validation_message <- renderUI({
+      validate(
+        need(length(selected_items()) > 0,
+             HTML("<div style='color: #d9534f;'>No files selected</div>")
+        )
+      )
     })
+
+    output$main_panel_dynamic <- renderUI({
+      req(selected_items(), members)
+      tryCatch({
+        if (length(selected_items()) == 0) {
+          return(HTML("<div style='font-size: small !important; font-family: \"Helvetica Neue\", Helvetica, Arial, sans-serif !important; color: #a94442; font-weight: 700;'>No files selected (required)</div>"))
+        }
+
+        w_load_items$show()
+
+        log_string <- glue::glue_collapse(selected_items(), sep = ", ")
+        debug(.le$logger, glue::glue("Files selected for QC: {log_string}"))
+
+        relevant_files_list <- tryCatch({
+          relevant_files()
+        }, error = function(e){
+          NULL
+        })
+
+        list <- render_selected_list(
+          input = input,
+          ns = ns,
+          items = selected_items(),
+          checklist_choices = checklists,
+          relevant_files = relevant_files_list,
+          output = output,
+          members = members
+        )
+
+        isolate_rendered_list(input = input,
+                              session = session,
+                              items = selected_items(),
+                              members = members)
+
+        session$sendCustomMessage("adjust_grid", id) # finds the width of the files and adjusts grid column spacing based on values
+        return(list)
+      }, error = function(e) {
+        error(.le$logger, glue::glue("There was an error rendering items in right panel: {e$message}"))
+        stopApp()
+        rlang::abort(e$message)
+      })
+    })
+
 
     observe({
       req(input$adjust_grid_finished) # retrieve msg through js when adjust grid is done
+      req(selected_items())
+      items <- selected_items()
+      for (name in items) {
+        checklist_input_id <- generate_input_id("checklist", name)
+        debug(.le$logger, glue::glue("Adding validation rule for {checklist_input_id}"))
+        iv$add_rule(checklist_input_id, shinyvalidate::sv_required())
+      }
+
       w_load_items$hide()
     })
 
     observeEvent(selected_items(), {
+      req(checklists)
       items <- selected_items()
       for (name in items) {
         log_string <- glue::glue_collapse(items, sep = ", ")
@@ -239,9 +297,12 @@ return "<div><strong>" + escape(item.username) + "</div>"
         tryCatch(
           {
             create_button_preview_event(input, name = name)
+            associate_relevant_files_button_event(input = input, output = output, name = name, ns = ns, root_dir = root_dir, relevant_files = relevant_files)
+            create_checklist_preview_event(input = input, name = name, checklists = checklists)
           },
           error = function(e) {
             error(.le$logger, glue::glue("There was an error creating the preview buttons: {e$message}"))
+            stopApp()
             rlang::abort(e$message)
           }
         )
@@ -270,6 +331,7 @@ return "<div><strong>" + escape(item.username) + "</div>"
         },
         error = function(e) {
           error(.le$logger, glue::glue("There was an error retrieving one of the status_checks items: {e$message}"))
+          stopApp()
           rlang::abort(e$message)
         }
       )
@@ -334,12 +396,13 @@ return "<div><strong>" + escape(item.username) + "</div>"
                       files = qc_items()
           )
 
-          create_checklists("test.yaml")
+          create_checklists("test.yaml", remote)
           removeClass("create_qc_items", "enabled-btn")
           addClass("create_qc_items", "disabled-btn")
         },
         error = function(e) {
           error(.le$logger, glue::glue("There was an error creating the Milestone {qc_items()}: {e$message}"))
+          stopApp()
           rlang::abort(e$message)
         }
       )
@@ -434,7 +497,13 @@ return "<div><strong>" + escape(item.username) + "</div>"
       addClass("create_qc_items", "disabled-btn")
 
       if (length(selected_items()) > 0 && isTruthy(rv_milestone())) {
-        file_data <- extract_file_data(input, selected_items())
+        relevant_files_list <- tryCatch({
+          relevant_files()
+        }, error = function(e){
+          NULL
+        })
+
+        file_data <- extract_file_data(input, selected_items(), relevant_files_list)
         if (!is.null(file_data)) {
           debug(.le$logger, glue::glue("create_qc_items buttons are activated because there are {length(selected_items())} selected items and milestone is named {rv_milestone()}"))
           removeClass("create_qc_items", "disabled-btn")
@@ -462,26 +531,8 @@ return "<div><strong>" + escape(item.username) + "</div>"
 
     observeEvent(input$reset, {
       debug(.le$logger, glue::glue("App was reset through the reset button."))
+      reset_triggered(TRUE)
       session$reload()
-    })
-
-    #HERE
-    observeEvent(selected_items(), {
-      req(checklists)
-      items <- selected_items()
-      for (name in items) {
-        log_string <- glue::glue_collapse(items, sep = ", ")
-        debug(.le$logger, glue::glue("Preview buttons created for: {log_string}"))
-        tryCatch(
-          {
-            create_checklist_preview_event(input, iv, ns, name = name, checklists)
-          },
-          error = function(e) {
-            error(.le$logger, glue::glue("There was an error creating the preview buttons: {e$message}"))
-            rlang::abort(e$message)
-          }
-        )
-      }
     })
 
     iv$enable()