+++ R/format_commit_shas.R
@@ -7,6 +7,7 @@
 # add "second most recent commit", "most recent commit" and "original qc commit" identifiers
 # format in table
 
+
 get_commits_df <- function(issue_number, owner, repo, remote) {
   init_qc_commit <- get_init_qc_commit(owner, repo, issue_number)
   branch <- get_branch_from_metadata(owner, repo, issue_number)